The variables changed the type of data by size_t Commit

+++ 0-bubble_sort.c
@@ -8,7 +8,7 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	unsigned int i = 0, j = 0, temp;
+	size_t i = 0, j = 0, temp;
 
 	for ( ; i < size ; i++)
 	{
